client/python/pyproject.toml
[project]
name = "gravitino-demo-client"
version = "0.1.0"
description = "A client for demostrating interaction with Apache Gravitino"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apache-gravitino>=1.0.0",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "ipython>=9.7.0",
    "ruff>=0.14.4",
]
